requirements.txt
Pillow>=10.0.0
aiosqlite>=0.19.0
aiohttp>=3.8.0